Changed blank decal from alpha_tester to beta_tester
Since I used the alpha_tester decal for the roundel on the Carul de Lupta R35 Modernizat (and for the upcoming Romanian pack), I've switched the transparent decal to beta_tester, since none of my mods use that yet. Additionally, fewer vehicles use the alpha_tester decal than the beta_tester one, so I figured it'd make more sense to have the more common one be blank instead of a roundel for an unrelated nation to the vehicles to which the beta tester decal is applied to.